-Working NodeQueue implementation that is significantly faster than Java's PriorityQueue
-Proper best-case pathfinding if the destination cannot be reached (will select shortest path to location with the lowest possible f-value)

-Lots of minor tweaks and fixes

-PathNodes now store Double instead of int

+++ arena-api/src/main/java/io/github/zap/arenaapi/pathfind/DestinationSelector.java
@@ -4,11 +4,11 @@ import org.jetbrains.annotations.NotNull;
 
 public interface DestinationSelector {
     DestinationSelector CLOSEST = (operation, node) -> {
-        int bestDistance = Integer.MAX_VALUE;
+        double bestDistance = Double.MAX_VALUE;
         PathDestination bestDestination = null;
 
         for(PathDestination destination : operation.getDestinations()) {
-            int sample = node.distanceSquaredTo(destination.node());
+            double sample = node.distanceSquaredTo(destination.node());
 
             if(sample < bestDistance) {
                 bestDistance = sample;
