+++ arena-api/src/main/java/io/github/zap/arenaapi/util/VectorUtils.java
@@ -17,15 +17,15 @@ public final class VectorUtils {
                 Math.abs(first.getZ() - second.getZ());
     }
 
-    public static int distanceSquared(int x1, int y1, int z1, int x2, int y2, int z2) {
-        int xD = x1 - x2;
-        int yD = y1 - y2;
-        int zD = z1 - z2;
+    public static double distanceSquared(double x1, double y1, double z1, double x2, double y2, double z2) {
+        double xD = x1 - x2;
+        double yD = y1 - y2;
+        double zD = z1 - z2;
 
-        return  (xD * xD) + (yD * yD) + (zD * zD);
+        return (xD * xD) + (yD * yD) + (zD * zD);
     }
 
-    public static double distance(int x1, int y1, int z1, int x2, int y2, int z2) {
+    public static double distance(double x1, double y1, double z1, double x2, double y2, double z2) {
         return Math.sqrt(distanceSquared(x1, y1, z1, x2, y2, z2));
     }
